requirements.txt
APScheduler==3.11.0
fastapi==0.115.6
msgpack_python==0.5.6
pycryptodome==3.20.0
python-dotenv==1.0.1
Requests==2.32.3
starlette==0.41.3
uvicorn==0.32.1
